requirements.txt
asgiref==3.5.2
Django==3.2.13
django-admin-sortable2==1.0.4
django-grappelli==2.13.1
django-tinymce==3.5.0
djangorestframework==3.13.1
easy-thumbnails==2.7.1
mock==4.0.3
mysqlclient==2.0.3
Pillow==9.1.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==4.2.0
uWSGI==2.0.19.1
xlwt==1.3.0
